requirements.txt
# Azure Dependencies
azure-storage-blob>=12.19.0
opencensus-ext-azure>=1.1.13
opencensus-context>=0.1.3

# Translation & Processing
requests>=2.31.0
pyodbc>=5.0.1

# Data Processing
lxml>=4.9.3
